Refactor the disk integrations so they're a part of the trigger system lib.
