+++ trigger-system/src/iface_impl/config.rs
@@ -1,4 +1,7 @@
 use std::fmt;
+use std::fs;
+use std::io;
+use std::path::{Path, PathBuf};
 
 use gcloud::{datastore::DatastoreClient, AuthProvider};
 
@@ -7,6 +10,8 @@ use crate::interface::{TriggerConfigLoader, TriggerConfiguration};
 #[derive(Debug)]
 pub enum Error {
     DatastoreError(String),
+    SerializationError(serde_json::Error),
+    IOError(io::Error),
 }
 
 impl fmt::Display for Error {
@@ -38,5 +43,22 @@ impl TriggerConfigLoader for DatastoreTriggerConfigLoader {
             .get_all()
             .map_err(|ds| Error::DatastoreError(format!("{:?}", ds)))?;
         Ok(results)
+    }
+}
+
+/// Reads trigger configurations from a file.
+pub struct FileTriggerConfigLoader {
+    path: PathBuf,
+}
+
+impl TriggerConfigLoader for FileTriggerConfigLoader {
+    type Error = Error;
+
+    fn get_all_configurations(&self) -> Result<Vec<TriggerConfiguration>, Self::Error> {
+        let handle = fs::File::open(&self.path).map_err(Error::IOError)?;
+        let value: Vec<TriggerConfiguration> =
+            serde_json::from_reader(handle).map_err(Error::SerializationError)?;
+
+        Ok(value)
     }
 }
