+++ trigger-system/src/iface_impl/mod.rs
@@ -1,5 +1,5 @@
 mod config;
 mod trigger_writer;
 
-pub use config::DatastoreTriggerConfigLoader;
-pub use trigger_writer::PubsubTriggerQueueWriter;
+pub use config::{DatastoreTriggerConfigLoader, FileTriggerConfigLoader};
+pub use trigger_writer::{DirectoryTriggerQueueWriter, PubsubTriggerQueueWriter};
